src/jwt/JwtService.spec.ts
import * as jwt from 'jsonwebtoken';
import { Test } from '@nestjs/testing';

import { JwtService } from './JwtService';
import { CONFIG_OPTIONS } from '../shared/constants/CoreConstants'; // 이건 뭘까??

const TEST_KEY = 'testKey';
const PAY_LOAD = { id: 1 };

jest.mock('jsonwebtoken', () => {
	return {
		sign: jest.fn(() => 'TOKEN'),
		verify: jest.fn(() => PAY_LOAD),
	};
});

describe('JwtService', () => {
	let service: JwtService;

	beforeEach(async () => {
		const module = await Test.createTestingModule({
			providers: [
				JwtService,
				{
					provide: CONFIG_OPTIONS,
					useValue: { privateKey: TEST_KEY },
				},
			],
		}).compile();

		service = module.get<JwtService>(JwtService);
	});

	it('be defined', () => {
		expect(service).toBeDefined();
	});

	describe('sign', () => {
		it('should return a signed token', () => {
			const token = service.sign(PAY_LOAD);

			expect(jwt.sign).toHaveBeenCalledTimes(1);
			expect(jwt.sign).toHaveBeenCalledWith(PAY_LOAD, TEST_KEY);
			expect(typeof token).toBe('string');
		});
	});

	describe('verify', () => {
		it('should return the decoded token', () => {
			const TOKEN = 'TOKEN';
			const decodedToken = service.verify(TOKEN);

			expect(jwt.verify).toHaveBeenCalledTimes(1);
			expect(jwt.verify).toHaveBeenCalledWith(TOKEN, TEST_KEY);
			expect(decodedToken).toEqual(PAY_LOAD);
		});
	});
});
